Move critic_tools and critic_main to props/core/critic/
Consolidate critic-specific code under props/core/critic/:
- tools.py (was agent_loop/critic_tools.py)
- main.py (was agent_loop/critic_main.py)

agent_loop/ now only contains the generic loop.py module.
