+++ props/core/critic/tools.py
